fix(form): tweak operational hours form to correctly submit only dirty fields, fix closed checkbox
