+++ apps/web/src/components/dashboard/team/overview/summary-section.tsx
@@ -1,12 +1,7 @@
 'use client'
 
-import { useEffect, useMemo } from 'react'
+import { useMemo } from 'react'
 import { useParams, useRouter } from 'next/navigation'
-import {
-  OperationalHourOutput,
-  OperationalHourOutputSchema,
-} from '@fuku/api/schemas'
-import { DayOfWeekKey } from '@fuku/db/schemas'
 import {
   Badge,
   Button,
@@ -15,34 +10,12 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-  Checkbox,
-  Field,
-  FieldContent,
-  FieldError,
-  FieldLabel,
-  Item,
-  ItemActions,
-  ItemContent,
-  ItemTitle,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
   Separator,
 } from '@fuku/ui/components'
-import { zodResolver } from '@hookform/resolvers/zod'
 import { useMutation, useQueries, useQuery } from '@tanstack/react-query'
 import { ArrowRight, Calendar } from 'lucide-react'
-import {
-  Controller,
-  SubmitHandler,
-  useForm,
-} from 'react-hook-form'
-import { toast } from 'sonner'
 
 import { useSession } from '~/components/providers/session-provider'
-import { TIME_OPTIONS, WEEKDAY_MAP } from '~/lib/date'
 import { isEntity } from '~/lib/db'
 import { useTRPC } from '~/trpc/client'
 
@@ -324,213 +297,6 @@ export const SummarySection = () => {
         {shiftTypesSummary}
         {payGradeSummary}
       </div>
-      <OperationalHoursSection teamId={team?.id ?? ''} />
-    </div>
-  )
-}
-
-function OperationalHoursSection({ teamId }: { teamId: string }) {
-  const trpc = useTRPC()
-
-  const { data: operationalHours, isSuccess } = useQuery({
-    ...trpc.operationalHour.list.queryOptions({ teamId }),
-    enabled: !!teamId,
-  })
-
-  const { mutateAsync: setHours } = useMutation({
-    ...trpc.operationalHour.setHours.mutationOptions(),
-    onSuccess: () => {
-      toast.success('Operational Hours', {
-        description: 'Changes saved!',
-      })
-    },
-    onError: error => {
-      toast.error('Operational Hours', {
-        description: error.message,
-      })
-    },
-  })
-
-  const defaultDay = {
-    teamId,
-    startTime: '',
-    endTime: '',
-    deletedAt: null as Date | null,
-  }
-
-  const form = useForm<OperationalHourOutput>({
-    resolver: zodResolver(OperationalHourOutputSchema),
-    defaultValues: {
-      1: defaultDay,
-      2: defaultDay,
-      3: defaultDay,
-      4: defaultDay,
-      5: defaultDay,
-      6: defaultDay,
-      7: defaultDay,
-    },
-  })
-
-  const resetForm = () => {
-    if (operationalHours) {
-      form.reset(operationalHours, {
-        keepDirtyValues: false,
-        keepDirty: false,
-      })
-    }
-  }
-
-  const onSubmit: SubmitHandler<OperationalHourOutput> = async values => {
-    try {
-      const dirtyFields = form.formState.dirtyFields
-
-      // Extract days where ANY nested field changed
-      const dirtyDays = Object.keys(dirtyFields).filter(day => {
-        const value = dirtyFields[day as keyof typeof dirtyFields]
-        return value && Object.keys(value as object).length > 0
-      }) as DayOfWeekKey[]
-
-      if (dirtyDays.length === 0) {
-        form.setError('root', {
-          type: 'manual',
-          message: 'There are no changes to save.',
-        })
-        return
-      }
-
-      const hours = dirtyDays.reduce((acc, day) => {
-        acc[day] = values[day]
-        return acc
-      }, {} as OperationalHourOutput)
-
-      await setHours({ teamId, hours })
-    } catch {
-      // handled in mutation onError
-    }
-  }
-
-  useEffect(() => {
-    if (isSuccess && operationalHours) {
-      form.reset(operationalHours)
-    }
-  }, [isSuccess, operationalHours])
-
-  return (
-    <div className='flex flex-col gap-4'>
-      <h2>Operational Hours</h2>
-
-      <form
-        onSubmit={form.handleSubmit(onSubmit)}
-        className='flex flex-col gap-2'
-      >
-        {(Object.keys(WEEKDAY_MAP) as DayOfWeekKey[]).map(day => {
-          const weekday = WEEKDAY_MAP[day]
-          return (
-            <Item size='xs' key={day} className='flex items-center gap-4'>
-              <ItemContent>
-                <ItemTitle>{weekday}</ItemTitle>
-
-                <ItemActions>
-                  {/* CLOSED CHECKBOX */}
-                  <Controller
-                    control={form.control}
-                    name={`${day}.deletedAt`}
-                    render={({ field }) => (
-                      <Field orientation='horizontal'>
-                        <Checkbox
-                          id={`closed-${day}`}
-                          name={`closed-${day}`}
-                          disabled={!isSuccess}
-                          checked={field.value !== null}
-                          onCheckedChange={checked =>
-                            field.onChange(checked ? new Date() : null)
-                          }
-                        />
-                        <FieldContent>
-                          <FieldLabel htmlFor={`closed-${day}`}>
-                            Closed
-                          </FieldLabel>
-                        </FieldContent>
-                      </Field>
-                    )}
-                  />
-
-                  {/* START TIME */}
-                  <Controller
-                    control={form.control}
-                    name={`${day}.startTime`}
-                    render={({ field }) => (
-                      <Field orientation='horizontal'>
-                        <FieldLabel className='sr-only'>Start Time</FieldLabel>
-                        <Select
-                          disabled={!isSuccess}
-                          value={field.value}
-                          onValueChange={field.onChange}
-                        >
-                          <SelectTrigger className='w-full max-w-24'>
-                            <SelectValue placeholder='-' />
-                          </SelectTrigger>
-                          <SelectContent className='w-full max-w-24'>
-                            {TIME_OPTIONS.map(time => (
-                              <SelectItem key={time} value={time}>
-                                {time}
-                              </SelectItem>
-                            ))}
-                          </SelectContent>
-                        </Select>
-                      </Field>
-                    )}
-                  />
-
-                  {/* END TIME */}
-                  <Controller
-                    control={form.control}
-                    name={`${day}.endTime`}
-                    render={({ field }) => (
-                      <Field orientation='horizontal'>
-                        <FieldLabel className='sr-only'>End Time</FieldLabel>
-                        <Select
-                          disabled={!isSuccess}
-                          value={field.value}
-                          onValueChange={field.onChange}
-                        >
-                          <SelectTrigger className='w-full max-w-24'>
-                            <SelectValue placeholder='-' />
-                          </SelectTrigger>
-                          <SelectContent className='w-full max-w-24'>
-                            {TIME_OPTIONS.map(time => (
-                              <SelectItem key={time} value={time}>
-                                {time}
-                              </SelectItem>
-                            ))}
-                          </SelectContent>
-                        </Select>
-                      </Field>
-                    )}
-                  />
-                </ItemActions>
-              </ItemContent>
-            </Item>
-          )
-        })}
-
-        <Field orientation='horizontal' className='gap-4 last:mt-2'>
-          <Button type='submit' disabled={!isSuccess}>
-            Save Changes
-          </Button>
-
-          <Button
-            type='button'
-            disabled={!isSuccess}
-            onClick={resetForm}
-            variant='outline'
-          >
-            Reset
-          </Button>
-
-          <FieldError>{form.formState.errors.root?.message}</FieldError>
-        </Field>
-      </form>
     </div>
   )
 }
